eightqueens/eightqueens_test.go
package eightqueens

import (
	"fmt"
	"math/rand"
	"reflect"
	"testing"
)

func TestRandomOrganism(t *testing.T) {
	for i := 0; i < 50; i++ {
		t.Run(fmt.Sprintf("Random Organism %v", i), func(t *testing.T) {
			got := RandomOrganism(8)
			if len(got) != 8 {
				t.Errorf("RandomOrganism() = %v, want 8", len(got))
			}
			for g := range got {
				if got[g] < 1 || got[g] > 8 {
					t.Errorf("RandomOrganism() didn't limit correctly: got: %v", got[g])
				}
			}
		})
	}
}

func TestFitness(t *testing.T) {
	type args struct {
		organism Organism
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{
			name: "BAD",
			args: args{
				organism: Organism([]int{1, 1, 1, 1, 1, 1, 1, 1}),
			},
			want: 21,
		},
		// {
		// 	name: "Worse",
		// 	args: args{
		// 		organism: Organism([]int{1, 2, 3, 4, 5, 6, 7, 8}),
		// 	},
		// 	want: 20, // this isn't working, it returns 28 ?!
		// },
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := Fitness(tt.args.organism); got != tt.want {
				t.Errorf("Fitness() = %v, want %v", got, tt.want)
			}
		})
	}
}

func TestReproduce(t *testing.T) {
	rand.Seed(1)
	type args struct {
		x Organism
		y Organism
	}
	tests := []struct {
		name string
		args args
		want Organism
	}{
		{
			name: "simple",
			args: args{
				x: Organism([]int{1, 1, 1, 1, 1, 1, 1, 1}),
				y: Organism([]int{7, 7, 7, 7, 7, 7, 7, 7}),
			},
			want: Organism([]int{1, 7, 7, 7, 7, 7, 7, 7}),
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := Reproduce(tt.args.x, tt.args.y); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("Reproduce() = %v, want %v", got, tt.want)
			}
		})
	}
}

func TestMutate(t *testing.T) {
	rand.Seed(1)
	type args struct {
		x Organism
	}
	tests := []struct {
		name string
		args args
		want Organism
	}{
		{name: "mutate 1", args: args{x: Organism([]int{1, 1, 1, 1, 1, 1, 1, 1})}, want: Organism([]int{1, 8, 1, 1, 1, 1, 1, 1})},
		{name: "mutate 2", args: args{x: Organism([]int{1, 1, 1, 1, 1, 1, 1, 1})}, want: Organism([]int{1, 1, 1, 1, 1, 1, 1, 4})},
		{name: "mutate 3", args: args{x: Organism([]int{1, 1, 1, 1, 1, 1, 1, 1})}, want: Organism([]int{1, 7, 1, 1, 1, 1, 1, 1})},
		{name: "mutate 4", args: args{x: Organism([]int{1, 1, 1, 1, 1, 1, 1, 1})}, want: Organism([]int{1, 5, 1, 1, 1, 1, 1, 1})},
		{name: "mutate 5", args: args{x: Organism([]int{1, 1, 1, 1, 1, 1, 1, 1})}, want: Organism([]int{5, 1, 1, 1, 1, 1, 1, 1})},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := Mutate(tt.args.x); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("Mutate() = %v, want %v", got, tt.want)
			}
		})
	}
}

func TestRandomSelection(t *testing.T) {
	rand.Seed(1)
	type args struct {
		population    []Organism
		probabilities []float64
	}
	tests := []struct {
		name string
		args args
		want Organism
	}{
		{
			name: "random 1",
			args: args{
				population: []Organism{
					Organism([]int{1, 1, 1, 1, 1, 1, 1, 1}),
					Organism([]int{2, 2, 2, 2, 2, 2, 2, 2}),
					Organism([]int{3, 3, 3, 3, 3, 3, 3, 3}),
					Organism([]int{4, 4, 4, 4, 4, 4, 4, 4}),
				},
				probabilities: []float64{0.5, 0.2, 0.3, 0.4},
			},
			want: Organism{2, 2, 2, 2, 2, 2, 2, 2},
		},
		{
			name: "random 2",
			args: args{
				population: []Organism{
					Organism([]int{1, 1, 1, 1, 1, 1, 1, 1}),
					Organism([]int{2, 2, 2, 2, 2, 2, 2, 2}),
					Organism([]int{3, 3, 3, 3, 3, 3, 3, 3}),
					Organism([]int{4, 4, 4, 4, 4, 4, 4, 4}),
				},
				probabilities: []float64{0.1, 0.2, 0.6, 0.4},
			},
			want: Organism{4, 4, 4, 4, 4, 4, 4, 4},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := RandomSelection(tt.args.population, tt.args.probabilities); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("RandomSelection() = %v, want %v", got, tt.want)
			}
		})
	}
}
